backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
aiofiles==24.1.0
websockets==12.0
python-dotenv==1.0.1
Pillow==10.4.0
openpyxl==3.1.5
httpx==0.28.1
psycopg2-binary==2.9.11
